requirements.txt
requests
click
numpy
bs4
pypdf2
tabula-py
openpyxl
